pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aigis"
version = "2.0.0"
description = "The open-source firewall for AI agents. Block prompt injections, jailbreaks, and data leaks before they reach your LLM. Multi-layer defense, agent-era security (MCP/Capability), US/CN/JP/EU compliance. Zero-dependency core."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = [
    "llm", "security", "prompt-injection", "ai-safety",
    "pii", "jailbreak", "guardrails", "owasp",
    "ai-agent", "governance", "middleware", "fastapi",
    "langchain", "openai", "anthropic", "mcp",
    "mcp-security", "red-team", "mitre-atlas",
    "ai-firewall", "agent-security", "aigis",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
]
# Zero required dependencies for the core library
dependencies = []

[project.optional-dependencies]
# YAML-based policy files
yaml = ["pyyaml>=6.0"]

# FastAPI / Starlette middleware
fastapi = ["fastapi>=0.115.0", "starlette>=0.41.0"]

# LangChain callback
langchain = ["langchain-core>=0.1.0"]

# OpenAI proxy wrapper
openai = ["openai>=1.0.0"]

# Anthropic SDK integration
anthropic = ["anthropic>=0.20.0"]

# Full SaaS server (backend + all integrations)
server = [
    "aigis[yaml,fastapi,openai]",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "stripe>=8.0.0",
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
]

# Everything
all = ["aigis[yaml,fastapi,langchain,openai,anthropic]"]

# Development
dev = [
    "aigis[all]",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
aigis = "aigis.cli:main"

[project.urls]
Homepage = "https://github.com/killertcell428/aigis"
Documentation = "https://github.com/killertcell428/aigis#readme"
Issues = "https://github.com/killertcell428/aigis/issues"
Changelog = "https://github.com/killertcell428/aigis/releases"

[tool.hatch.build.targets.wheel]
packages = ["aigis"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
per-file-ignores = { "aigis/filters/patterns.py" = ["E501"], "aigis/patterns.py" = ["E501"], "aigis/similarity.py" = ["E501"], "aigis/badge.py" = ["E501"], "aigis/cli.py" = ["E501", "F401"], "aigis/activity.py" = ["E501"], "aigis/compliance.py" = ["E501"], "aigis/scanner.py" = ["E501"], "aigis/multi_agent/message_scanner.py" = ["E501"], "aigis/cross_session/sleeper.py" = ["E501"], "aigis/cross_session/correlator.py" = ["E501"], "tests/*.py" = ["E501"] }

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
disallow_any_generics = false
